Rename workspace/existing-file action to workspace/existing-file-with-content
