+++ documentation/installation.md
@@ -48,8 +48,9 @@ configure TextRunner, and run:
 $ text-runner setup
 </pre>
 
-This creates a configuration file called <a type="workspace/existing-file">
-**text-runner.jsonc** that looks like this:
+This creates a configuration file called
+<a type="workspace/existing-file-with-content"> **text-runner.jsonc** with this
+content:
 
 ```jsonc
 {
