+++ textrun-repo/README.md
@@ -16,17 +16,15 @@ To use these actions, add this package as a development dependency by running
 ### repo/existing-file
 
 Sometimes you want to mention the name of a file from your application's source
-code in the documentation. The
-<b type="action/name-full">source/existing-file</b> action verifies this. As an
-example, consider a codebase contains a file
-<a type="workspace/new-file">**config.yml** with content `foo: bar`</a>. The
-documentation for this codebase might want to mention this configuration file:
+code in the documentation. The <b type="action/name-full">repo/existing-file</b>
+action verifies this. As an example, consider a codebase contains a file
+<a type="workspace/empty-file">**config.yml**</a>. The documentation for this
+codebase might want to mention this configuration file:
 
 <a type="extension/runnable-region">
 
 ```markdown
-The file <b type="repo/existing-file">config.yml</b> defines configuration
-values.
+File <b type="repo/existing-file">config.yml</b> defines configuration values.
 ```
 
 </a>
