+++ textrun-workspace/src/actions/existing-file.ts
@@ -1,18 +1,15 @@
-import * as assertNoDiff from "assert-no-diff"
 import * as color from "colorette"
 import { promises as fs } from "fs"
 import * as path from "path"
 import * as textRunner from "text-runner-core"
 
 export async function existingFile(action: textRunner.actions.Args): Promise<void> {
-  const fileName = action.region.textInNodeOfType("strong", "em")
+  const fileName = action.region.text()
   const fileRelPath = path.join(action.region[0].attributes["dir"] || ".", fileName)
-  action.name(`verify content of file ${color.cyan(fileRelPath)}`)
+  action.name(`verify existence of file ${color.cyan(fileRelPath)}`)
   const fullPath = action.configuration.workspace.joinStr(fileRelPath)
-  const expectedContent = action.region.textInNodeOfType("fence", "code")
-  action.log(expectedContent)
   try {
-    var actualContent = await fs.readFile(fullPath, "utf-8")
+    await fs.stat(fullPath)
   } catch (e) {
     if (!textRunner.isFsError(e)) {
       throw e
@@ -26,13 +23,5 @@ export async function existingFile(action: textRunner.actions.Args): Promise<voi
     } else {
       throw e
     }
-  }
-  try {
-    assertNoDiff.trimmedLines(actualContent.trim(), expectedContent.trim())
-  } catch (err) {
-    throw new textRunner.UserError(
-      `mismatching content in ${color.cyan(color.bold(fileRelPath))}`,
-      textRunner.errorMessage(err)
-    )
   }
 }
