+++ textrun-workspace/src/index.ts
@@ -3,6 +3,7 @@ import * as textRunner from "text-runner"
 import { additionalFileContent } from "./actions/additional-file-content.js"
 import { emptyFile } from "./actions/empty-file.js"
 import { existingDirectory } from "./actions/existing-directory.js"
+import { existingFileWithContent } from "./actions/existing-file-with-content.js"
 import { existingFile } from "./actions/existing-file.js"
 import { newDirectory } from "./actions/new-directory.js"
 import { newFile } from "./actions/new-file.js"
@@ -13,6 +14,7 @@ export const textrunActions: textRunner.exports.TextrunActions = {
   emptyFile,
   existingDirectory,
   existingFile,
+  existingFileWithContent,
   newDirectory,
   newFile,
   workingDir
